requirements.txt
Flask==3.0.2
spotipy==2.25.1
python-dotenv==1.1.0
beautifulsoup4==4.13.4